EXPLORING THE POWER OF VISUALIZATION
One thing I have to come to understand is in machine learning and data science coding takes approximately 20% of the data science and machine learning process. The other percentage is taken by your thinking processes, math, domain expertise, and much more.

An example of the thought process, I am working on a supermarket till transactions dataset which turned out to be interesting. In the month of April, the highest spend was on Monday, but this amount was paid in the morning and this was the pay of the highest quantity that month at it was shopped at 3 pm. 

One customer, say,  Customer  A was the 3rd highest spender per hour, fourth overall highest spender, bought the highest quantity in April, shops only at the store in region W01, shops at 3 pm, shops on Friday and Tuesday. They, however, shopped heavily on Friday as compared to Tuesday.